Refactor event names
