--- src/interfaces/events.ts
@@ -1,3 +1,0 @@
-export type GetBookmarksEvent = "getBookmarks";
-export type RefreshBookmarksEvent = "refreshBookmarks";
-export type OpenApiTokenWebsiteEvent = "openApiTokenWebsite";
